refactor: split diagnostics module
