+++ pkgs/standards/autoapi/autoapi/v3/system/diagnostics/__init__.py
@@ -1,0 +1,23 @@
+from ...runtime.kernel import build_phase_chains
+from .router import mount_diagnostics
+from .methodz import build_methodz_endpoint as _build_methodz_endpoint
+from .hookz import build_hookz_endpoint as _build_hookz_endpoint
+from .planz import build_planz_endpoint as _build_planz_endpoint
+from .utils import (
+    model_iter as _model_iter,
+    opspecs as _opspecs,
+    label_callable as _label_callable,
+    label_hook as _label_hook,
+)
+
+__all__ = [
+    "mount_diagnostics",
+    "_build_methodz_endpoint",
+    "_build_hookz_endpoint",
+    "_build_planz_endpoint",
+    "_model_iter",
+    "_opspecs",
+    "_label_callable",
+    "_label_hook",
+    "build_phase_chains",
+]
